Maybe reuse existing escaping
