+++ packages/core/src/common/shell-quoting.spec.ts
